refactor: move `src/test/schema` directory to `src/schema`
