+++ src/test/cpp/index.ts
@@ -1,8 +1,8 @@
 import path from "node:path";
 import { compileCppSource } from "../../compile/cpp.js";
 import { runExecutable } from "../../run.js";
-import { parseCppTestSchema } from "../schema/cpp.js";
-import { readRawTestSchema } from "../schema/raw.js";
+import { parseCppTestSchema } from "../../schema/cpp.js";
+import { readRawTestSchema } from "../../schema/raw.js";
 import { generateCppTest } from "./generate/index.js";
 
 /**
